Refactor data access & update Cover entity
This commit encompasses a series of changes aimed at enhancing the data model and access patterns within the application. Firstly, the data access method in Program.cs was refactored from `RetrieveAnArtistWithTheirCovers()` to `GetAnAuthorWithTheirBooksWithCoversWithArtists()`, significantly expanding its functionality to fetch and display comprehensive details about authors, their books, and the associated covers and artists.

In PubContext.cs, the `Cover` entity was updated to include a `BookId`, establishing a direct one-to-one relationship with the `Book` entity. This modification is part of moving towards a more normalized data model, ensuring covers are directly linked to books.

The model snapshot in PubContextModelSnapshot.cs was adjusted to reflect the new `BookId` column in the `Covers` table, setting up a unique index to enforce the one-to-one relationship between books and covers. This adjustment includes configuring a cascade delete behavior for the relationship.

Furthermore, navigation properties were added to the `Book` and `Cover` classes in Book.cs and Cover.cs, respectively. These changes facilitate easier and more intuitive data access patterns within the application by directly linking books to their covers and vice versa.

Lastly, a new migration, 20240626203045_covertobook.cs, along with its corresponding designer file, was added to implement the database schema changes necessary for the one-to-one relationship between the `Book` and `Cover` entities. This includes adding the `BookId` column to the `Covers` table, updating existing data, creating a unique index on `BookId`, and setting up a foreign key constraint.